OneDrive/Desktop/Pr3/task3_face_blur/requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
numpy
torch  
